Implement incremental mode
If two are passed on the command line rather than one, only the items
present in the second file but not the first will be modified.

This is intended to support incremental update of Anki with only new
Pocket items, which can run more often than non-incremental update and
sync less data each time, in order to make Anki easier to use on a
frequent basis without have to wait for a long sync every few hours.