Add Pocket Article notes without item_id to Pocket
Pocket has the disadvantage that it doesn't support any kind of personal
notes on its items. This is a major weakness if I'm bookmarking stuff
I read on the web, because I can't preserve any personal context to
remind myself later why I cared about the item enough to save it in the
first place. Creating new notes in Anki circumvents this limitation,
because my Anki cards have a `personal_notes` field.

Now, notes with `given_url` and `given_title` set, but without any
`item_id` (which is the item's ID in the Pocket API) are added to Pocket
using the API, before processing any items. This ensures that these
newly-added Pocket items can be processed by the rest of the script as
if they had been present all along in the Pockexport data in the file
passed to the script.